+++ src/fetchers/fetchPullRequestById/index.js
@@ -1,3 +1,5 @@
+const parsePullRequest = require('./parser');
+
 const PR_BY_ID_QUERY = `
   query($id: ID!) {
     node(id: $id) {
@@ -6,6 +8,14 @@ const PR_BY_ID_QUERY = `
         url
         body
         number
+        comments(last: 100) {
+          nodes {
+            author {
+              login
+            }
+            body
+          }
+        }
       }
     }
   }
@@ -15,6 +25,7 @@ module.exports = (octokit, id) => {
   const variables = { id };
   return octokit
     .graphql(PR_BY_ID_QUERY, variables)
+    .then(parsePullRequest)
     .catch((error) => {
       const msg = `Error fetching pull requests with id "${id}"`;
       throw new Error(`${msg}. Error: ${error}`);