+++ src/fetchers/fetchPullRequestById/parser.js
@@ -1,0 +1,10 @@
+const get = require('lodash.get');
+
+const parseComments = (node) => ({
+  ...node,
+});
+
+module.exports = ({ node: data }) => ({
+  ...data,
+  comments: (get(data, 'comments.nodes') || []).map(parseComments),
+});