+++ src/fetchers/index.js
@@ -1,3 +1,4 @@
+const commentOnPullRequest = require('./commentOnPullRequest');
 const fetchPullRequestById = require('./fetchPullRequestById');
 const fetchPullRequests = require('./fetchPullRequests');
 const fetchSponsorships = require('./fetchSponsorships');
@@ -5,6 +6,7 @@ const postToSlack = require('./postToSlack');
 const updatePullRequest = require('./updatePullRequest');
 
 module.exports = {
+  commentOnPullRequest,
   fetchPullRequestById,
   fetchPullRequests,
   fetchSponsorships,