+++ src/interactors/__tests__/alreadyPublished.test.js
@@ -4,6 +4,10 @@ const STATS = '## Pull reviewers stats\n|stats|table|';
 const OTHER_CONTENT = '## Other pull request content';
 
 describe('Interactors | .alreadyPublished', () => {
+  it('returns false when input is falsy', () => {
+    expect(alreadyPublished(null)).toBe(false);
+  });
+
   it('returns false when body is empty', () => {
     const body = '';
     const pullRequest = { body };
@@ -25,6 +29,16 @@ describe('Interactors | .alreadyPublished', () => {
   it('returns true when body contains other stuff and stats', () => {
     const body = `${OTHER_CONTENT}\n${STATS}`;
     const pullRequest = { body };
+    expect(alreadyPublished(pullRequest)).toBe(true);
+  });
+
+  it('returns true when a comment contains stats', () => {
+    const comments = [
+      {
+        body: STATS,
+      },
+    ];
+    const pullRequest = { body: '', comments };
     expect(alreadyPublished(pullRequest)).toBe(true);
   });
 });