+++ src/interactors/index.js
@@ -2,7 +2,6 @@ const alreadyPublished = require('./alreadyPublished');
 const buildTable = require('./buildTable');
 const buildComment = require('./buildComment');
 const checkSponsorship = require('./checkSponsorship');
-const getPullRequest = require('./getPullRequest');
 const getPulls = require('./getPulls');
 const getReviewers = require('./getReviewers');
 const postSlackMessage = require('./postSlackMessage');
@@ -14,7 +13,6 @@ module.exports = {
   buildTable,
   buildComment,
   checkSponsorship,
-  getPullRequest,
   getPulls,
   getReviewers,
   postSlackMessage,